MGM: add hardlink support

+++ mgm/FuseServer.cc
@@ -23,6 +23,9 @@
 
 /*----------------------------------------------------------------------------*/
 
+#include <string>
+#include <cstdlib>
+
 #include "mgm/FuseServer.hh"
 #include "mgm/Acl.hh"
 #include "mgm/Policy.hh"
@@ -40,6 +43,10 @@ EOSMGMNAMESPACE_BEGIN
 #define SA_OK 64   // set xattr
 #define U_OK 128   // can update
 #define SU_OK 256  // set utime
+
+const char *k_mdino = "sys.eos.mdino";
+const char *k_nlink = "sys.eos.nlink";
+
 /*----------------------------------------------------------------------------*/
 FuseServer::FuseServer()
 {
@@ -1568,6 +1575,8 @@ FuseServer::FillFileMD(uint64_t inode, eos::fusex::md & file)
   // fills file meta data by inode number
 
   eos::FileMD* fmd = 0;
+  eos::FileMD* gmd = 0;
+
   eos::FileMD::ctime_t ctime;
   eos::FileMD::ctime_t mtime;
   uint64_t clock=0;
@@ -1576,11 +1585,23 @@ FuseServer::FillFileMD(uint64_t inode, eos::fusex::md & file)
 
   try
   {
+    bool has_mdino = false;
     fmd = gOFS->eosFileService->getFileMD(eos::common::FileId::InodeToFid(inode), &clock);
     eos_static_info("clock=%llx", clock);
+
+    file.set_name(fmd->getName());
+    gmd = fmd;
+    if (fmd->hasAttribute(k_mdino)) {
+      has_mdino = true;
+      uint64_t mdino = std::stoll(fmd->getAttribute(k_mdino));
+      fmd = gOFS->eosFileService->getFileMD(eos::common::FileId::InodeToFid(mdino), &clock);
+      eos_static_info("hlnk switched from %s to file %s (%#llx)", gmd->getName().c_str(), fmd->getName().c_str(), mdino);
+    }
+    
+    /* fmd = link target file, gmd = link file */
     fmd->getCTime(ctime);
     fmd->getMTime(mtime);
-    file.set_md_ino(inode);
+    file.set_md_ino(eos::common::FileId::FidToInode(gmd->getId()));
     file.set_md_pino(fmd->getContainerId());
     file.set_ctime(ctime.tv_sec);
     file.set_ctime_ns(ctime.tv_nsec);
@@ -1603,14 +1624,21 @@ FuseServer::FillFileMD(uint64_t inode, eos::fusex::md & file)
     {
       file.set_mode(fmd->getFlags() | S_IFREG );
     }
-    // TODO: no hardlinks
-    file.set_nlink(1);
-    file.set_name(fmd->getName());
+
+    /* hardlinks */
+    int nlink = 0;
+    if (fmd->hasAttribute(k_nlink)) {
+      nlink = std::stoi(fmd->getAttribute(k_nlink));
+      eos_static_debug("hlnk %s (%#lx) nlink %d", file.name().c_str(), fmd->getId(), nlink);
+    }
+    file.set_nlink(nlink);
+
     file.set_clock(clock);
 
     for ( eos::FileMD::XAttrMap::iterator it = fmd->attributesBegin();
          it != fmd->attributesEnd(); ++it)
     {
+      if (has_mdino && ((it->first) == k_nlink)) continue;
       (*file.mutable_attr())[it->first] = it->second;
       if ( (it->first) == "sys.eos.btime")
       {
@@ -1620,6 +1648,10 @@ FuseServer::FillFileMD(uint64_t inode, eos::fusex::md & file)
         file.set_btime_ns(strtoul(val.c_str(), 0, 10));
       }
     }
+
+    if (has_mdino) {
+      (*file.mutable_attr())[k_mdino] = gmd->getAttribute(k_mdino);
+    }    
 
     file.clear_err();
 
@@ -2679,43 +2711,92 @@ FuseServer::HandleMD(const std::string &id,
                           (long) fid,
                           (long) md.md_ino(),
                           (long) md.md_pino(), (long) fmd->getContainerId());
-        }
-        else
-        {
-          // file creation
-          op = CREATE;
-
-          unsigned long layoutId = 0;
-          unsigned long forcedFsId = 0;
-          long forcedGroup = 0;
-          XrdOucString space;
-          eos::ContainerMD::XAttrMap attrmap = pcmd->getAttributeMap();
-          XrdOucEnv env;
-
-          // retrieve the layout
-          Policy::GetLayoutAndSpace("fusex",
-                                    attrmap,
-                                    *vid,
-                                    layoutId,
-                                    space,
-                                    env,
-                                    forcedFsId,
-                                    forcedGroup);
-
-
-	  if (fmd)
+        } 
+	else 
+	{
+	  if (md.target().substr(0,8) == "////hlnk" ) 
 	  {
-	    eos_static_crit("discovered re-creation of existing file");
+	    // creation of a hard link
+	    uint64_t tgt_md_ino = atoll(md.target().c_str()+8);
+	    
+	    pcmd = gOFS->eosDirectoryService->getContainerMD(md.md_pino());
+	    if (pcmd->findContainer(md.name()))                /* name check protected by eosViewRWMutex above */
+	      return EEXIST;
+	    
+	    /* fmd is the target file corresponding to tgt_fid, gmd the file corresponding to new name */
+	    fmd = gOFS->eosFileService->getFileMD(eos::common::FileId::InodeToFid(tgt_md_ino));
+	    eos::FileMD* gmd = gOFS->eosFileService->createFile();
+	    
+	    int nlink;                                  /* here: 0 origin, 1 means file + one hard link */
+	    nlink = (fmd->hasAttribute(k_nlink)) ? std::stoi(fmd->getAttribute(k_nlink)) : 0;
+	    eos_static_debug("hlnk fid=%#lx target name %s nlink %d create hard link %s",
+			     (long) fid, fmd->getName().c_str(), nlink, md.name().c_str());
+	    nlink += 1;
+	    fmd->setAttribute(k_nlink, std::to_string(nlink));
+	    gOFS->eosFileService->updateStore(fmd);
+	    
+	    gmd->setAttribute(k_mdino, std::to_string(tgt_md_ino));
+	    gmd->setName(md.name());
+	    gOFS->eosFileService->updateStore(gmd);
+	    
+	    eos_static_debug("hlnk %s mdino %s %s nlink %s", gmd->getName().c_str(), gmd->getAttribute(k_mdino).c_str(),
+			     fmd->getName().c_str(), fmd->getAttribute(k_nlink).c_str());
+	    
+	    pcmd->addFile(gmd);
+	    gOFS->eosView->updateContainerStore(pcmd);
+	    
+	    eos::fusex::response resp;
+	    resp.set_type(resp.ACK);
+	    resp.mutable_ack_()->set_code(resp.ack_().OK);
+	    resp.mutable_ack_()->set_transactionid(md.reqid());
+	    resp.mutable_ack_()->set_md_ino(eos::common::FileId::FidToInode(gmd->getId()));
+	    resp.SerializeToString(response);
+	    struct timespec pt_mtime;
+	    pt_mtime.tv_sec = md.mtime();
+	    pt_mtime.tv_nsec = md.mtime_ns();
+	    
+	    gOFS->eosDirectoryService->updateStore(pcmd);
+	    uint64_t clock = 0;
+	    Cap().BroadcastMD(md, tgt_md_ino, md_pino, clock, pt_mtime);
+	    return 0;
 	  }
-
-	  fmd = gOFS->eosFileService->createFile();
-	 
-          fmd->setName( md.name() );
-          fmd->setLayoutId( layoutId );
-          md_ino = eos::common::FileId::FidToInode(fmd->getId());
-          pcmd->addFile(fmd);
-          eos_static_info("ino=%lx pino=%lx md-ino=%lx create-file", (long) md.md_ino(), (long) md.md_pino(), md_ino);
-        }
+	  else
+	  {
+	    // file creation
+	    op = CREATE;
+	    
+	    unsigned long layoutId = 0;
+	    unsigned long forcedFsId = 0;
+	    long forcedGroup = 0;
+	    XrdOucString space;
+	    eos::ContainerMD::XAttrMap attrmap = pcmd->getAttributeMap();
+	    XrdOucEnv env;
+	    
+	    // retrieve the layout
+	    Policy::GetLayoutAndSpace("fusex",
+				      attrmap,
+				      *vid,
+				      layoutId,
+				      space,
+				      env,
+				      forcedFsId,
+				      forcedGroup);
+	    
+	    
+	    if (fmd)
+	      {
+		eos_static_crit("discovered re-creation of existing file");
+	      }
+	    
+	    fmd = gOFS->eosFileService->createFile();
+	    
+	    fmd->setName( md.name() );
+	    fmd->setLayoutId( layoutId );
+	    md_ino = eos::common::FileId::FidToInode(fmd->getId());
+	    pcmd->addFile(fmd);
+	    eos_static_info("ino=%lx pino=%lx md-ino=%lx create-file", (long) md.md_ino(), (long) md.md_pino(), md_ino);
+	  }
+	}
 
         fmd->setName(md.name());
         fmd->setCUid(md.uid());
@@ -3029,9 +3110,67 @@ FuseServer::HandleMD(const std::string &id,
 
         }
 
-        pcmd->removeFile(fmd->getName());
-        fmd->setContainerId(0);
-        fmd->unlinkAllLocations();
+	bool doDelete = true;
+	uint64_t tgt_md_ino;
+	
+	if (fmd->hasAttribute(k_mdino)) 
+        {
+	  // this is a hard link, update reference count on underlying file 
+	  tgt_md_ino = std::stoll(fmd->getAttribute(k_mdino));
+	  uint64_t clock;
+	  
+	  // gmd = the file holding the inode 
+	  eos::FileMD* gmd = gOFS->eosFileService->getFileMD(eos::common::FileId::InodeToFid(tgt_md_ino), &clock);
+	  long nlink = std::stol(gmd->getAttribute(k_nlink)) - 1;
+	  
+	  if (nlink >= 0) 
+          {
+	    gmd->setAttribute(k_nlink, std::to_string(nlink));
+	    gOFS->eosFileService->updateStore(gmd);
+	    eos_static_info("hlnk nlink update on %s for %s now %ld", gmd->getName().c_str(), fmd->getName().c_str(), nlink);
+	  } 
+          else 
+          {              
+	    // remove target file as well
+	    eos_static_info("hlnk unlink target %s for %s nlink %ld", gmd->getName().c_str(), fmd->getName().c_str(), nlink);
+	    pcmd->removeFile(gmd->getName());
+	    gmd->setContainerId(0);
+	    gmd->unlinkAllLocations();
+	    gOFS->eosFileService->updateStore(gmd);
+	  }
+	} 
+	else 
+	{
+	  if (fmd->hasAttribute(k_nlink)) 
+	  {    
+	    // this is a genuine file, potentially with hard links
+	    tgt_md_ino = eos::common::FileId::FidToInode(fmd->getId());
+	    // reduce reference count, only remove file if negative (origin 0 == 1 file) 
+	    long nlink = std::stol(fmd->getAttribute(k_nlink)) - 1;
+	    if (nlink >= 0) 
+	    {  
+	      // hard links exist, just rename the file so the inode does not disappear
+	      char nameBuf[256];
+	      snprintf(nameBuf, sizeof(nameBuf), "...eos.ino...%lx", fmd->getId());
+	      std::string tmpName = nameBuf;
+	    
+	      fmd->setAttribute(k_nlink, std::to_string(nlink));
+	      eos_static_info("hlnk unlink rename %s=>%s new nlink %d", fmd->getName().c_str(), tmpName.c_str(), nlink);
+	      pcmd->removeFile(tmpName);         // if the target exists, remove it!
+	      gOFS->eosView->renameFile(fmd, tmpName);
+	      doDelete = false;
+	    } else
+	      eos_static_info("hlnk nlink %ld for %s, will be deleted", nlink, fmd->getName().c_str());
+	  }
+	}
+
+	if (doDelete) {
+	  pcmd->removeFile(fmd->getName());
+	  fmd->setContainerId(0);
+	  fmd->unlinkAllLocations();
+	}
+
+
         gOFS->eosFileService->updateStore(fmd);
         gOFS->eosDirectoryService->updateStore(pcmd);
         pcmd->notifyMTimeChange( gOFS->eosDirectoryService );
